move out hdfs tests
