+++ hdfs-integration-test/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "hdfs-integration-test"
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+readme.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dev-dependencies]
+delta_kernel = { path = "../kernel", features = ["arrow-55", "default-engine"] }
+# must be compatible with arrow version (and hdfs-native)
+hdfs-native-object-store = { version = "0.14.0" }
+hdfs-native = "0.11.1"
+walkdir = { version = "2.5.0" }
+tokio = { version = "1.44.0", features = ["full"] }
+url = "2.4.0"
+
+[features]
+integration-test = ["hdfs-native-object-store/integration-test"]
